requirements.txt
Flask
beautifulsoup4
requests
openai
tmdbv3api
python-dotenv
cloudscraper
